requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Pillow==10.4.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10
Werkzeug==3.1.3
SQLAlchemy==2.0.41